chore: extract aws to its own package
