+++ packages/js-plugin-aws/babel.config.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../babel.config.json"
+}
