+++ packages/js-plugin-aws/rollup.config.js
@@ -1,0 +1,16 @@
+import { createConfigs } from '../../rollup.config';
+import pkg from './package.json';
+
+export default createConfigs({
+  pkg,
+  builds: [
+    {
+      dir: 'dist/esm',
+      format: 'es',
+    },
+    {
+      dir: 'dist/cjs',
+      format: 'cjs',
+    },
+  ],
+});
