+++ packages/js-plugin-aws/src/index.ts
