+++ packages/js-plugin-aws/src/plugin.ts
